+++ tests/util/test_geometry.py
@@ -733,6 +733,10 @@ class TestBoundingBox:
         assert BoundingBox(1, 2, 3, 4).is_georeferenced() is False
         assert BoundingBox(1, 2, 3, 4, crs=4326).is_georeferenced() is True
 
+    def test_has_crs(self):
+        assert BoundingBox(1, 2, 3, 4).has_crs() is False
+        assert BoundingBox(1, 2, 3, 4, crs=4326).has_crs() is True
+
     def test_as_tuple(self):
         bbox = BoundingBox(1, 2, 3, 4)
         assert bbox.as_tuple() == (1, 2, 3, 4, None)
@@ -758,6 +762,15 @@ class TestBoundingBox:
             "type": "Polygon",
             "coordinates": (((3, 2), (3, 4), (1, 4), (1, 2), (3, 2)),),
         }
+
+    def test_approx(self):
+        expected = BoundingBox(100, 200, 300, 400)
+        bbox = BoundingBox(100.01, 200.01, 300.01, 400.01)
+        assert bbox == expected.approx(abs=0.1)
+        assert bbox != expected.approx(abs=0.001)
+
+        assert bbox == expected.approx(rel=0.001)
+        assert bbox != expected.approx(rel=0.0001)
 
     def test_as_geojson_from_utm(self):
         bbox = BoundingBox(500000, 5649824, 507016, 5660950, crs="EPSG:32631")
@@ -825,6 +838,132 @@ class TestBoundingBox:
 
         # TODO: instead of this nonsense result, should this fail, or result in some special empty case?
         assert bbox.buffer(-10).as_tuple() == (11, 12, -7, -6, "EPSG:4326")
+
+    @pytest.mark.parametrize(
+        ["bbox", "target", "expected"],
+        [
+            (BoundingBox(11, 12, 13, 14), None, BoundingBox(11, 12, 13, 14, crs=None)),
+            (BoundingBox(11, 12, 13, 14, crs=4326), 4326, BoundingBox(11, 12, 13, 14, crs=4326)),
+            (BoundingBox(11, 12, 13, 14, crs=4326), "EPSG:4326", BoundingBox(11, 12, 13, 14, crs=4326)),
+            (
+                BoundingBox(11, 12, 13, 14, crs=4326),
+                None,
+                BoundingBoxException("Can not reproject bounding box with CRS 'EPSG:4326' to unspecified CRS."),
+            ),
+            (
+                BoundingBox(11, 12, 13, 14, crs=None),
+                "EPSG:4326",
+                CrsRequired("Can not reproject bounding box with unspecified CRS to 'EPSG:4326'."),
+            ),
+            (
+                BoundingBox(11, 12, 13, 14, crs=4326),
+                BoundingBox(5, 6, 7, 8, crs=4326),
+                BoundingBox(11, 12, 13, 14, crs=4326),
+            ),
+            (
+                BoundingBox(3.1, 51.1, 3.2, 51.2, crs=4326),
+                32631,
+                BoundingBox(west=506986, south=5660950, east=514003, north=5672085, crs="EPSG:32631").approx(abs=1),
+            ),
+        ],
+    )
+    def test_align_to(self, bbox, target, expected):
+        if isinstance(expected, BoundingBox):
+            aligned = bbox.align_to(target)
+            assert aligned == expected
+        elif isinstance(expected, Exception):
+            with pytest.raises(type(expected), match=str(expected)):
+                _ = bbox.align_to(target)
+        else:
+            raise ValueError(expected)
+
+    @pytest.mark.parametrize(
+        ["bbox1", "bbox2", "expected"],
+        [
+            (
+                BoundingBox(11, 12, 13, 14),
+                BoundingBox(16, 15, 17, 18),
+                BoundingBox(11, 12, 17, 18),
+            ),
+            (
+                BoundingBox(11, 12, 13, 14, crs=4326),
+                BoundingBox(16, 15, 17, 18, crs=4326),
+                BoundingBox(11, 12, 17, 18, crs=4326),
+            ),
+            (
+                BoundingBox(3.1, 51.1, 3.11, 51.11, crs="epsg:4326"),
+                BoundingBox(500000, 5649824, 507016, 5660950, crs="EPSG:32631"),
+                BoundingBox(3, 51, 3.11, 51.11, crs="EPSG:4326").approx(abs=1e-4),
+            ),
+            (
+                BoundingBox(500000, 5649824, 507016, 5660950, crs="EPSG:32631"),
+                BoundingBox(3.1, 51.1, 3.11, 51.11, crs="epsg:4326"),
+                BoundingBox(500000, 5649824, 507702, 5662063, crs="EPSG:32631").approx(abs=1),
+            ),
+        ],
+    )
+    def test_union(self, bbox1, bbox2, expected):
+        result = bbox1.union(bbox2)
+        assert result == expected
+
+    @pytest.mark.parametrize(
+        ["bbox1", "bbox2", "expected"],
+        [
+            (
+                # Simple overlap
+                BoundingBox(1, 2, 5, 6),
+                BoundingBox(3, 4, 8, 9),
+                BoundingBox(3, 4, 5, 6),
+            ),
+            (
+                # large x-shift
+                BoundingBox(1, 2, 5, 6),
+                BoundingBox(13, 4, 18, 9),
+                None,
+            ),
+            (
+                # Large y-shift
+                BoundingBox(1, 2, 5, 6),
+                BoundingBox(3, 14, 8, 19),
+                None,
+            ),
+            (
+                # One larger containing the other
+                BoundingBox(10, 20, 30, 40),
+                BoundingBox(12, 22, 14, 24),
+                BoundingBox(12, 22, 14, 24),
+            ),
+            (
+                # explicit CRS
+                BoundingBox(1, 2, 5, 6, crs=4326),
+                BoundingBox(3, 4, 8, 9, crs="EPSG:4326"),
+                BoundingBox(3, 4, 5, 6, crs="EPSG:4326"),
+            ),
+            (
+                BoundingBox(3.1, 51.1, 3.3, 51.3, crs="epsg:4326"),
+                # Roughly (3.2, 51.2, 3.4, 51.4) in lot-lan:
+                BoundingBox(513912, 5672085, 527946, 5694383, crs="EPSG:32631"),
+                BoundingBox(3.2, 51.2, 3.3, 51.3, crs="EPSG:4326").approx(abs=0.001),
+            ),
+            (
+                # Roughly (3.2, 51.2, 3.4, 51.4) in lot-lan:
+                BoundingBox(513912, 5672085, 527946, 5694383, crs="EPSG:32631"),
+                BoundingBox(3.1, 51.1, 3.3, 51.3, crs="epsg:4326"),
+                BoundingBox(513912, 5672085, 521005, 5683229, crs="EPSG:32631").approx(abs=1),
+            ),
+            (
+                BoundingBox(13.1, 51.1, 13.3, 51.3, crs="epsg:4326"),
+                BoundingBox(513912, 5672085, 527946, 5694383, crs="EPSG:32631"),
+                None,
+            ),
+        ],
+    )
+    def test_intersection(self, bbox1, bbox2, expected):
+        assert bbox1.intersection(bbox2) == expected
+
+        # When there is no reprojection in play we can easily verify the reverse case too
+        if bbox1.crs == bbox2.crs:
+            assert bbox2.intersection(bbox1) == expected
 
 
 class TestValidateGeoJSON: